backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
face_recognition
opencv-python
Pillow
numpy
pandas
python-dotenv
pytest
pytest-django
pytest-cov
gunicorn
dj-database-url
whitenoise
psycopg2-binary
djangorestframework-csv